tests/wpt/web-platform-tests/document-picture-in-picture/requires-secure-context.html

<!DOCTYPE html>
<title>Test that documentPictureInPicture is not available in insecure contexts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
test(() => {
  assert_equals(undefined, window.documentPictureInPicture);
},
    "documentPictureInPicture should not be available in insecure contexts");
</script>
</body>
